Move metadata related logic
